auth

+++ src/main/java/codeholic/controller/MemberController.java
@@ -1,25 +1,35 @@
 package codeholic.controller;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import codeholic.domain.Member;
 import codeholic.domain.Response;
 import codeholic.domain.request.RequestLoginUser;
+import codeholic.exception.AlreadyLogoutException;
 import codeholic.service.AuthService;
 import codeholic.service.CookieUtil;
 import codeholic.service.JwtUtil;
 import codeholic.service.RedisUtil;
 import io.jsonwebtoken.ExpiredJwtException;
-//
+
 @RestController
 @CrossOrigin(origins="*")
 @RequestMapping("user")
@@ -29,8 +39,6 @@ public class MemberController {
                     "\"admin\",\"domain\": { \"id\": \"default\" },\"password\": \"cert0188!\"}}},\"scope\": {\"project\": {\"name\": \"admin\",\"domain\": { \"id\": \"default\" }}}}}";
 
     public final static String user_create_d ="{\"user\": {\"name\": \"newuser\", \"password\": \"changeme\"}}";
-    
-
 
     @Autowired
     private AuthService authService;
@@ -45,14 +53,24 @@ public class MemberController {
     private RedisUtil redisUtil;
 
     @PostMapping("/signup")
-    public Response signUpUser(@RequestBody Member member){
+    public Response signUpUser(@Valid @RequestBody Member member){
         Response response = new Response();
+
+        //notblank 위반시 유효성 검사, 나중에 다른 코드로 빼기
+        
+        //실행
         try{
             authService.signUpUser(member);
             response.setResponse("success");
             response.setMessage("회원가입을 성공적으로 완료했습니다.");
         }
+        catch(IllegalStateException ie){
+            response.setResponse("failed");
+            response.setMessage("회원가입을 하는 도중 오류가 발생했습니다.");
+            response.setData(ie.getMessage());
+        }
         catch(Exception e){
+            //아이디 중복시 오류검사....
             response.setResponse("failed");
             response.setMessage("회원가입을 하는 도중 오류가 발생했습니다.");
             response.setData(e.toString());
@@ -60,8 +78,9 @@ public class MemberController {
 
         return response;
     }
+    
     @PostMapping("/signin")
-    public Response login(@RequestBody RequestLoginUser user,
+    public Response login(@Valid @RequestBody RequestLoginUser user,
                           HttpServletRequest req,
                           HttpServletResponse res) {
         try {
@@ -86,22 +105,51 @@ public class MemberController {
     public Response logout(
                             HttpServletRequest req,
                             HttpServletResponse res) {
+        Response response = new Response();
         String username;
-        String accessToken = req.getHeader("Authorization").substring(7); //bearer 표기
+        String accessToken = null;
         try{
+            
+            accessToken = req.getHeader("Authorization").substring(7);
             username = jwtUtil.getUsername(accessToken);
+            //이미 로그아웃한 경우
+            if(redisUtil.getData(accessToken) == accessToken){
+                throw new AlreadyLogoutException();
+            }
             //access token을 블랙리스트로 올리기
             redisUtil.setDataExpire(accessToken, accessToken,  JwtUtil.REFRESH_TOKEN_VALIDATION_SECOND);
             //refresh token을 redis에서 삭제
             redisUtil.deleteData(username);
-        }catch(ExpiredJwtException e){
-            username = e.getClaims().getSubject();
+            response.setResponse("success");
+            response.setMessage("로그아웃에 성공했습니다.");
+        }catch(ExpiredJwtException ee){
+            username = ee.getClaims().getSubject();
             redisUtil.deleteData(username);
+        }catch(AlreadyLogoutException ae){
+            response.setResponse("failed");
+            response.setMessage("이미 로그아웃하였습니다.");
         }catch(Exception e){
-
+            response.setResponse("failed");
+            response.setMessage("로그아웃에 실패하였습니다.");
         }
-        return new Response("success", "로그아웃에 성공했습니다.", null);
+        return response;
 
+    }
+    // @Valid 에 대한 예외 처리 담당
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public Response handleValidationExceptions(MethodArgumentNotValidException ex) {
+        Map<String, String> errors = new HashMap<>();
+        ex.getBindingResult().getAllErrors().forEach((error) -> {
+            String fieldName = ((FieldError) error).getField();
+            String errorMessage = error.getDefaultMessage();
+             errors.put(fieldName, errorMessage);
+        });
+        Response response = new Response();
+        response.setData(errors);
+        response.setResponse("fail");
+        response.setMessage("오류가 발생했습니다.");
+        return response;
     }
 
 }
